% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_with_siren.R
\name{get_with_siren}
\alias{get_with_siren}
\title{Get with siren number}
\usage{
get_with_siren(siren, date = NULL, token)
}
\arguments{
\item{siren}{a SIREN Number}

\item{date}{the date to start from}

\item{token}{your API token}
}
\value{
a list of data.frame
}
\description{
Get with siren number
}
\examples{
dontrun{
get_with_siren(siren = "810064519",
    date = "2017-01-01",
    token = token)
}

}
