% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_with_siret.R
\name{get_with_siret}
\alias{get_with_siret}
\title{Get with siret}
\usage{
get_with_siret(siret, date = NULL, token)
}
\arguments{
\item{siret}{a siret number}

\item{date}{the date to start from}

\item{token}{an api token}
}
\value{
a list of data.frame
}
\description{
Get with siret
}
\examples{
dontrun{
get_with_siret(siret = "810064519",
    date = "2017-01-01",
    token = token)
}

}
