% tinyrox says don't edit this manually, but it can't stop you!
\name{gpu_cache_stats}
\alias{gpu_cache_stats}
\title{Get GPU Kernel Cache Statistics}
\usage{
gpu_cache_stats()
}
\value{
List with n_cached (number of kernels in cache) and
  kernel_names (character vector of cached kernel names)
}
\description{
Get GPU Kernel Cache Statistics
}
