% tinyrox says don't edit this manually, but it can't stop you!
\name{triton_op_supported}
\alias{triton_op_supported}
\title{Check if an Op is Supported for Triton Emission}
\usage{
triton_op_supported(op)
}
\arguments{
\item{op}{Character, operation name from torchlang IR}
}
\value{
Logical
}
\description{
Check if an Op is Supported for Triton Emission
}
